Split reviews index for easier refactoring
