+++ examples/demo/src/App.js
@@ -26,7 +26,7 @@ import {
 } from './products';
 import invoices from './invoices';
 import { CategoryList, CategoryEdit, CategoryIcon } from './categories';
-import { ReviewList, ReviewEdit, ReviewIcon } from './reviews';
+import reviews from './reviews';
 
 import dataProviderFactory from './dataProvider';
 import fakeServerFactory from './fakeServer';
@@ -113,12 +113,7 @@ class App extends Component {
                     edit={CategoryEdit}
                     icon={CategoryIcon}
                 />
-                <Resource
-                    name="reviews"
-                    list={ReviewList}
-                    edit={ReviewEdit}
-                    icon={ReviewIcon}
-                />
+                <Resource name="reviews" {...reviews} />
             </Admin>
         );
     }
